refactor: 뉴스리스트를 레이아웃에서 패치
상세 페이지에서 즉시 활용하기 위한 목적
