+++ src/app/[locale]/news/page.tsx
@@ -1,11 +1,5 @@
-import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 import Container from '../news/__container';
-import { newsQueryOptions } from '@/features/news';
 import type { Metadata, ResolvingMetadata } from 'next';
-
-type Props = {
-  params: Promise<{ locale: string }>;
-};
 
 export async function generateMetadata(
   _props: unknown,
@@ -18,17 +12,6 @@ export async function generateMetadata(
   };
 }
 
-export default async function NewsPage({ params }: Props) {
-  const { locale } = await params;
-  const queryClient = new QueryClient();
-
-  await queryClient.prefetchQuery(newsQueryOptions(locale));
-
-  const dehydratedState = dehydrate(queryClient);
-
-  return (
-    <HydrationBoundary state={dehydratedState}>
-      <Container />
-    </HydrationBoundary>
-  );
+export default async function NewsPage() {
+  return <Container />;
 }
